move Config ..REG code to  config.h
